backend/requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
psycopg[binary]==3.2.3
supabase==2.7.4
